Feature/sign service

+++ src/main/java/repick/repickserver/global/config/JwtProperties.java
@@ -1,9 +1,7 @@
 package repick.repickserver.global.config;
 
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -12,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Data
 public class JwtProperties {
     private String secret;
-    private int accessTokenExpirationTime;
-    private int refreshTokenExpirationTime;
+    private Long accessTokenExpirationTime;
+    private Long refreshTokenExpirationTime;
     private String authoritiesKey;
 }
